attempt to resolve file problems with commits
